publish/serverservice: move server attribute methods into serverservice_attributes.go
This change is mainly for readability and an upcoming change
